+++ examples/examples/hyper_client.rs
@@ -25,8 +25,9 @@ use tower::{Service, ServiceBuilder, ServiceExt as _};
 
 use hitbox::Config;
 use hitbox::policy::PolicyConfig;
-use hitbox_http::extractors::Method as MethodExtractor;
+use hitbox_http::extractors::{MethodConfig, MethodExtractor};
 use hitbox_http::predicates::{NeutralRequestPredicate, NeutralResponsePredicate};
+use hitbox_http::request;
 use hitbox_moka::MokaBackend;
 use hitbox_tower::Cache;
 
@@ -48,7 +49,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     let config = Config::builder()
         .request_predicate(NeutralRequestPredicate::new())
         .response_predicate(NeutralResponsePredicate::new())
-        .extractor(MethodExtractor::new())
+        .extractor(request::extractor().method(MethodConfig::new()))
         .policy(PolicyConfig::builder().ttl(Duration::from_secs(60)).build())
         .build();