+++ examples/examples/reqwest.rs
@@ -25,8 +25,9 @@ use std::time::Duration;
 use hitbox::Config;
 use hitbox::policy::PolicyConfig;
 use hitbox_http::{
-    extractors::{Method as MethodExtractor, path::PathExtractor},
-    predicates::{NeutralResponsePredicate, request::Method as RequestMethod},
+    extractors::{MethodConfig, MethodExtractor, PathExtractor},
+    predicates::request::MethodPredicate,
+    request, response,
 };
 use hitbox_moka::MokaBackend;
 use hitbox_reqwest::CacheMiddleware;
@@ -48,9 +49,13 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // - Cache key includes: method + path
     // - TTL: 60 seconds
     let config = Config::builder()
-        .request_predicate(RequestMethod::new(http::Method::GET).unwrap())
-        .response_predicate(NeutralResponsePredicate::new())
-        .extractor(MethodExtractor::new().path("/{path}*"))
+        .request_predicate(request::predicate().method(http::Method::GET))
+        .response_predicate(response::predicate())
+        .extractor(
+            request::extractor()
+                .method(MethodConfig::new())
+                .path("/{path}*"),
+        )
         .policy(PolicyConfig::builder().ttl(Duration::from_secs(60)).build())
         .build();