+++ hitbox-configuration/src/predicates/response/status.rs
@@ -1,6 +1,7 @@
 use std::num::NonZeroU16;
 
-use hitbox_http::predicates::response::{StatusClass, StatusCode, StatusCodePredicate};
+use hitbox_http::predicates::response::status::Operation as StatusOp;
+use hitbox_http::predicates::response::{StatusClass, StatusCodePredicate};
 use http::StatusCode as HttpStatusCode;
 use hyper::body::Body as HttpBody;
 use serde::{Deserialize, Serialize};
@@ -116,18 +117,18 @@ impl Operation {
         match self {
             Operation::Eq(eq) => {
                 let status = parse_status_code(eq.status().get())?;
-                Ok(Box::new(inner.status_code(status)))
+                Ok(Box::new(inner.status(StatusOp::eq(status))))
             }
             Operation::In(r#in) => {
                 let status_codes = parse_status_codes(r#in.statuses())?;
-                Ok(Box::new(StatusCode::new_in(inner, status_codes)))
+                Ok(Box::new(inner.status(StatusOp::any(status_codes))))
             }
             Operation::Range { range } => {
                 let start = parse_status_code(range.start().get())?;
                 let end = parse_status_code(range.end().get())?;
-                Ok(Box::new(StatusCode::new_range(inner, start, end)))
+                Ok(Box::new(inner.status(StatusOp::range(start, end))))
             }
-            Operation::Class(class) => Ok(Box::new(StatusCode::new_class(inner, class.class()))),
+            Operation::Class(class) => Ok(Box::new(inner.status(StatusOp::class(class.class())))),
         }
     }
 }