+++ hitbox-tower/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Updated doc examples to hitbox-http 0.3 Config-based API ([#202](https://github.com/hit-box/hitbox/pull/202))
+
 ## [0.2.0] - 2026-01-27
 ### Added
 - Initial release